+++ Pester.psm1
@@ -1183,3 +1183,6 @@ Throw "This command has been renamed to 'Assert-VerifiableMock' (without the 's'
 & $script:SafeCommands['Export-ModuleMember'] SafeGetCommand, New-PesterOption
 & $script:SafeCommands['Export-ModuleMember'] Invoke-Gherkin, Find-GherkinStep, BeforeEachFeature, BeforeEachScenario, AfterEachFeature, AfterEachScenario, GherkinStep -Alias Given, When, Then, And, But
 & $script:SafeCommands['Export-ModuleMember'] New-MockObject, Add-AssertionOperator
+& $script:SafeCommands['Export-ModuleMember'] Add-PesterSuppression, Get-PesterSuppression, Clear-PesterSuppression
+
+$Script:PesterSuppression = @()